+++ hdf/src/dfufp2i.c
@@ -374,14 +374,14 @@ convert_interp(Input * in, Output * out)
     yoffs = (int32 *) HDmalloc((size_t) (out->vres + 1) * sizeof(int32));
     yoffs[0] = 0;
 
-    if (range < (float)0.0)
+    if (range < 0.0f)
         range = -range;     /* max must be > min */
 
     f = dys;    /* beginning of dys to fill in */
     yv = in->vscale;    /* beginning and end of yvals */
     lim = in->vscale + in->vdim - 2;
 
-    if (yrange > (float)0.0)
+    if (yrange > 0.0f)
       {
           for (i = 0; i < out->vres; i++)
             {   /* fill in dy's */
@@ -420,7 +420,7 @@ convert_interp(Input * in, Output * out)
     xv = in->hscale;    /* beginning and end of xvals */
     lim = in->hscale + in->hdim - 2;
 
-    if (xrange > (float)0.0)
+    if (xrange > 0.0f)
       {
           for (i = 0; i < out->hres; i++)
             {   /* fill in dx's */
@@ -482,7 +482,7 @@ convert_interp(Input * in, Output * out)
                 z = (*z1 - *z3 - *z2 + *z4) * (*xv) * zy +  /* weighted sum */
                     (*z3 - *z4) * (*xv) + (*z2 - *z4) * zy + *z4;
 
-                theval = (int) ((float)1.0 + (float)237.9 * (z - in->min) / range);   /* scaled value  */
+                theval = (int) (1.0f + 237.9f * (z - in->min) / range);   /* scaled value  */
                 if (theval >= 240 || theval < 1)
                     *p++ = 0;
                 else
@@ -535,7 +535,7 @@ pixrep_scaled(Input * in, Output * out)
     image = (uint8 *) out->image;   /* space for image */
 
     range = in->max - in->min;
-    if (range < (float)0.0)
+    if (range < 0.0f)
         range = -range;     /* max must be > min */
 
     hoffsets = (int32 *) HDmalloc((uint32) (out->hres + 1) * sizeof(int32));
@@ -545,7 +545,7 @@ pixrep_scaled(Input * in, Output * out)
     compute_offsets(in->vscale, in->vdim, voffsets, out->vres);
 
     prevoffset = voffsets[0] - 1;
-    ratio = (float) 237.9 / range;
+    ratio = 237.9f / range;
 
     for (i = 0; i < out->vres; i++)
       {     /* for each row, store pixel vals */
@@ -555,7 +555,7 @@ pixrep_scaled(Input * in, Output * out)
 
                 for (j = 0; j < in->hdim; j++)
                   {     /* compute vals for each data point */
-                      theval = (int) ((float)1.5 + ratio * (*data++ - in->min));
+                      theval = (int) (1.5f + ratio * (*data++ - in->min));
                       if (theval >= 240 || theval < 1)
                           theval = 0;
                       pixvals[j] = (uint8) theval;
@@ -609,7 +609,7 @@ compute_offsets(float *scale, int32 dim, int32 *offsets, int32 res)
 
     for (i = 0; i < dim - 1; i++)
       {     /* compute all midpoints */
-          midpt[i] = (scale[i] + scale[i + 1]) / (float) 2.0;
+          midpt[i] = (scale[i] + scale[i + 1]) / 2.0f;
 /*        printf("midpt[%d]=%8.1f\tscale[%d]=%8.1f\n",i,midpt[i],i,scale[i]); */
       }
     midpt[i] = scale[i] + scale[i] - midpt[i - 1];  /* tack one onto end */
@@ -656,7 +656,7 @@ pixrep_simple(Input * in, Output * out)
     float    *in_row_ptr, *in_buf;
     float     ratio, delh, delv, hblockend, vblockend;
 
-    ratio = (float) 237.9 / (in->max - in->min);
+    ratio = 237.9f / (in->max - in->min);
     image = (uint8 *) out->image;
     in_buf = in->data;
 
@@ -680,7 +680,7 @@ pixrep_simple(Input * in, Output * out)
             {
 
                 raster_val = (uint8)
-                    ((float)1.5 + ratio * (float) (*in_row_ptr++ - in->min));
+                    (1.5f + ratio * (float) (*in_row_ptr++ - in->min));
                 *image++ = raster_val;
 
                 for (; j < (int32) hblockend - 1; j++)  /* store vals for this blk of this row */
