+++ hdf/test/tvattr.c
@@ -62,7 +62,7 @@
 #define ATTNAME9   "attname9"
 #define ATTNAME10  "attname10"
 #define EPS64     (double)1.0E-14
-#define EPS32     (float)1.0E-7
+#define EPS32     1.0E-7f
 #define MAX_HDF4_NAME_LENGTH 256
 
 int32 data1[6]={0,-1,10,11,20,21}, idata1[6];
@@ -70,7 +70,7 @@ char  data2[6] = {'A','B','C','D','E','F'}, idata2[6];
 uint16 attr2[2] = {16, 32}, iattr2[2];
 uint32 attr1[4] = {100, 132, 10032, 10064}, iattr1[2];
 char   attr3[6] = {'m','N','p', 'S', 't', '\0'}, iattr3[6];
-float attr4[2] = {(float)32.001, (float)-34.002}, iattr4[2];
+float attr4[2] = {32.001f, -34.002f}, iattr4[2];
 double attr5[2] = {64.12345, -64.12345}, iattr5[2];
 
 static intn create_vset_stuff(void);
