+++ hdf/test/tvsfpack.c
@@ -79,8 +79,8 @@ fpack(void)
     pntr = databuf;
     /* pack a record at a time */
     for (i = 0; i < NRECORDS; i++) {
-        source[i].temp = (float)1.11 * (float)(i+1);
-        source[i].height = (float)2.22 * (float)(i+1);
+        source[i].temp = 1.11f * (float)(i+1);
+        source[i].height = 2.22f * (float)(i+1);
         source[i].speed = (int16)i;
         source[i].ident = (char)('A' + i);
          /* test error checks  */ 
@@ -134,8 +134,8 @@ fpack(void)
     /* pack a field at a time */
     pntr = databuf;
     for (i = 0; i < NRECORDS; i++) {
-        source[i].temp = (float)3.33 * (float)(i+1);
-        source[i].height = (float)4.44 * (float)(i+1);
+        source[i].temp = 3.33f * (float)(i+1);
+        source[i].height = 4.44f * (float)(i+1);
         source[i].speed = (int16)(2*i);
         source[i].ident = (char)('a' + i);
 	databufptr[0] = &source[i].speed;
@@ -230,8 +230,8 @@ funpack(void)
     }
 
     for (i=0; i<NRECORDS; i++)  
-         if ((iident[i] != (char)('A'+i)) || (fabs((double)(itemp[i] - (float)1.11*(float)(i+1))) > EPS) ||
-            (ispeed[i] != i) || (fabs((double)(iheight[i] - (float)2.22*(float)(i+1))) > EPS))  {
+         if ((iident[i] != (char)('A'+i)) || (fabs((double)(itemp[i] - 1.11f*(float)(i+1))) > EPS) ||
+            (ispeed[i] != i) || (fabs((double)(iheight[i] - 2.22f*(float)(i+1))) > EPS))  {
             num_errs++;
             printf(">>> Wrong data1 after VSfpack.\n");
         }
@@ -258,8 +258,8 @@ funpack(void)
     }
  
     for (i=0; i<NRECORDS; i++)  
-         if ((iident[i] != (char)('a'+i)) || (fabs((double)(itemp[i] - (float)3.33*(float)(i+1))) > EPS) ||
-            (ispeed[i] != 2*i) || (fabs((double)(iheight[i] - (float)4.44*(float)(i+1))) > EPS))  {
+         if ((iident[i] != (char)('a'+i)) || (fabs((double)(itemp[i] - 3.33f*(float)(i+1))) > EPS) ||
+            (ispeed[i] != 2*i) || (fabs((double)(iheight[i] - 4.44f*(float)(i+1))) > EPS))  {
             num_errs++;
             printf(">>> Wrong data2 after VSfpack.\n");
         }
@@ -292,7 +292,7 @@ funpack(void)
     }
     for (i=0; i<NRECORDS; i++)
          if ((iident[i] != 'A'+i) || (ispeed[i] != i) || 
-              (fabs((double)(iheight[i] - (float)2.22*(float)(i+1))) > EPS))  {
+              (fabs((double)(iheight[i] - 2.22f*(float)(i+1))) > EPS))  {
             num_errs++;
             printf(">>> Wrong subset data1 after VSfpack.\n");
         }
@@ -311,7 +311,7 @@ funpack(void)
     }
     for (i=0; i<NRECORDS; i++)
          if ((iident[i] != (char)('a'+i)) || (ispeed[i] != 2*i) || 
-              (fabs((double)(iheight[i] - (float)4.44*(float)(i+1))) > EPS)) {
+              (fabs((double)(iheight[i] - 4.44f*(float)(i+1))) > EPS)) {
             num_errs++;
             printf(">>> Wrong subset data2 after VSfpack.\n");
         }
